move staff related components into pageForStaff folder
